Remove the `SmfBuffer` type and improve docs
Added three `Smf::parse` variants in replacement of the `SmfBuffer`
type.

Made the `Smf<T>` type default to `Smf<Vec<Event>>`.

Renamed the `EventKind::read` public method to `EventKind::parse`.

Made parallelism optional through the `multithread` feature and
disabled it for the `TrackIter` representation, since the use of
multiple threads for a task as simple as creating an iterator
(virtually free) was more of a burden than a speedup.