Don't read the entire file in pathological cases
If the file was for example only zeroes, the parser would read the
entire file thinking it was 0-sized chunks with ID "\0\0\0\0", looking
for a header somewhere within the file.

Now the MIDI header MUST appear at the start of the file, or else a more
informative error is emitted as soon as the first 4 bytes are read.

Refactored, renamed, improved docs.